// admin/src/api/generated/admin.ts
export interface components {
    schemas: {
// ...
        HotlinkEntry: {
            hostname: string;
            enabled: boolean;
        };
        HotlinkSettingsView: {
            allowEmptyReferer: boolean;
            entries: components["schemas"]["HotlinkEntry"][];
        };
// ...
    };
// ...
}